Remove `package-lock.json`
The current version of this file in the template was created by someone using
Node v8.x, and one of the subdepenencies is broken for anyone using Node 10.x.
Deleting package-lock should make it work on either Node version, but most
importantly it should work on Node 10, which all the devs are now using.